Two-hand touch/pinch control
Pinch and hold with the left hand to move, or with both hands to
move/scale/rotY simultaneously

Scaling and translation are still not quite perfect: the way it should
work is that any point you touch becomes "stuck" to your hand, and its
movements perfectly match your hand movements (as long as the Y distance
between your hands remains constant).

The left-hand-only approach is temporary, I'll extend this to be
ambidextrous so that "left" and "right" are replaced with "first" and
"second".