Working (if slightly jarring) hand control
Pinch and hold with the left hand to move, or with both hands to
move/scale/rotY simultaneously

Scaling and translation are still not quite perfect: the way it should
work is that any point you touch becomes "stuck" to your hand, and its
movements perfectly match your hand movements (as long as the Y distance
between your hands remains constant).

The left-hand-only approach is temporary, I'll extend this to be
ambidextrous.

It could use some kind of double buffering to smooth the movement and
filter out hand tracking noise, which currently makes the whole scene a
jittery ~25fps mess while dragging

The correct approach should be:
        - Replace "left" and "right" with "first" and "second"
        - 1. Translate to put "first" at the origin
        - 2. Scale to match first-to-second distance
        - 3. Rotate to match position of "second"
        - 4. Translate back to put the camera in the right place