Better handling of loading spinners
If the preloader spinner is used, then it will disappear once the main
video player init() finishes. If multiple embedded players with
preloaders are used, their loading will not interfere with each other.

We should keep in mind that the whole idea of the preloader is only
because there's an annoyingly long load time before init() runs, and
there's probably a much better solution using dynamic import