Hold-touch-to-pause
Instead of the previous behavior where each touch toggles play/pause

This makes it easier to pause, pauses video during repositioning, and
prevents confusion about what state the video is in. It also means that
a false-positive touch event (which happens all the time for the Quest 3
in my room's lighting) only causes a momentary hiccup.